+++ doc/release-notes.md
@@ -4,4 +4,8 @@ Bitcoin ABC version 0.27.15 is now available from:
 
   <https://download.bitcoinabc.org/0.27.15/>
 
-This is a maintenance release with no user-visible change.
+This release includes the following features and fixes:
+ - The `getavalancheinfo` RPC has an improved `local.verification_status` field
+   that better explains why a proof is not verified. As a consequence the
+   `local.sharing` field is no longer needed and has been deprecated. You can
+   set the `-deprecatedrpc=getavalancheinfo_sharing` option to keep using it.